fix(#step2): 멀티 도메인 관련 기능 애플리케이션 서비스로 이동
